pulled abbreviate out into its own string.js file, can add other string manipulation functions there if necessary
